refactor(desktop): isolate YouTube link previews
